added a search to profile page and a search to the homepage

+++ api/src/routes/items/router.js
@@ -1,5 +1,5 @@
 const { Router } = require('express')
-const { showAllItems, getById, updateItem, deleteItem, addItem } = require('./controller')
+const { showAllItems, getById, updateItem, deleteItem, addItem, getByUserId, searchItems } = require('./controller')
 
 //import middleware
 const { authenticate } = require('../../middleware/auth')
@@ -8,9 +8,10 @@ const { authenticate } = require('../../middleware/auth')
 const router = new Router()
 
 // define routes
-
-router.get('/', showAllItems)
+router.get('/', showAllItems) // localhost:9000/items
+router.get('/get-items/:userId', getByUserId) // localhost:9000/items/get-items/1
 router.post('/', addItem)
+router.get('/search/:userId/:search', searchItems)
 router.get('/:itemId', getById)
 router.delete('/:itemId', deleteItem)
 router.patch('/:itemId', updateItem)
